+++ R/crew_options_metrics.R
@@ -1,12 +1,15 @@
 #' @title Options for logging resource usage metrics.
 #' @export
 #' @family options
-#' @description If a [crew_options_metrics()] object is
-#'   supplied to the `options_metrics` argument of a `crew`
-#'   controller function, then the `autometric` R package will
-#'   record resource usage metrics (such as CPU and memory usage)
-#'   for each running worker. Logging happens in
-#'   the background (through a detached POSIX) so as not to disrupt
+#' @description [crew_options_metrics()] configures the 
+#'   `crew` to record resource usage metrics (such as CPU and memory usage)
+#'   for each running worker.
+#'   To be activate resource usage logging,
+#'   the `autometric` R package version 0.1.0 or higher
+#'   must be installed.
+#'
+#'   Logging happens in the background (through a detached POSIX)
+#'   so as not to disrupt
 #'   the R session. On Unix-like systems, [crew_options_metrics()]
 #'   can specify `/dev/stdout` or `/dev/stderr` as the log files, which will
 #'   redirect output to existing logs you are already using.